+++ src/main/java/ruiseki/okbackpack/ClientProxy.java
@@ -10,7 +10,6 @@ import ruiseki.okbackpack.client.key.OpenBackpackHandler;
 import ruiseki.okbackpack.client.key.PickBlockHandler;
 import ruiseki.okbackpack.client.key.ToggleUpgradeHandler;
 import ruiseki.okbackpack.client.key.ToolSwapHandler;
-import ruiseki.okbackpack.client.tooltip.BackpackTooltipRenderer;
 import ruiseki.okbackpack.common.event.ItemRenderEvent;
 import ruiseki.okbackpack.compat.Mods;
 import ruiseki.okbackpack.compat.nei.BackpackGuiOpener;
@@ -77,6 +76,5 @@ public class ClientProxy extends ClientProxyComponent {
         super.registerEventHooks();;
         MinecraftForge.EVENT_BUS.register(new ItemRenderEvent());
         MinecraftForge.EVENT_BUS.register(BackpackJukeboxSoundManager.getInstance());
-        MinecraftForge.EVENT_BUS.register(new BackpackTooltipRenderer());
     }
 }
